Add renderer processor for render pipeline
