+++ docs/research/render_pipeline_composition_refactor.md
@@ -1,26 +1,23 @@
-# Render pipeline composition refactor note
+# Render Pipeline Composition Refactor
 
-## Problem statement
+## Problem Statement
 
-Document the render pipeline refactor that separates surface preparation, caching, and merge coordination so future contributors can reason about composition responsibilities and extend merge behavior safely.
+Describe the responsibilities of the new renderer-processing stage so the runtime pipeline keeps composition logic separate from per-renderer execution and logging.
 
 ## Materials
 
-- Local checkout of the Heart repository.
-- `src/heart/runtime/render_pipeline.py` for pipeline orchestration changes.
-- `src/heart/runtime/rendering/surface_provider.py` for surface preparation and caching.
-- `src/heart/runtime/rendering/surface_merge.py` for merge strategy selection and parallel merge coordination.
-- `docs/code_flow.md` for the updated runtime flow diagram and narrative.
+- Local checkout of this repository.
+- Python environment with the dependencies listed in `pyproject.toml` installed.
+- Source files for the renderer pipeline and rendering utilities.
 
-## Notes
+## Research Notes
 
-- Surface preparation now lives in a dedicated provider that owns display-mode enforcement and cached surface reuse, keeping `RenderPipeline` focused on orchestration.
-- Merge strategy selection and the parallel reduction loop are centralized in the composition manager so serial and parallel rendering share the same strategy controls.
-- `RenderPipeline.merge_surfaces` remains as the pairwise merge hook, allowing tests to override the merge function while using the shared composition manager.
+- `RendererProcessor` in `src/heart/runtime/rendering/renderer_processor.py` owns surface preparation, renderer initialization, internal processing, and per-renderer timing/logging so the pipeline can focus on orchestration.
+- `RenderPipeline` now delegates per-renderer work to `RendererProcessor` while retaining merge-strategy selection, composition, and parallel execution in `src/heart/runtime/render_pipeline.py`.
+- The flow diagram in `docs/code_flow.md` now highlights the processor as a distinct service between the pipeline and surface preparation.
 
-## Source references
+## Sources
 
+- `src/heart/runtime/rendering/renderer_processor.py`
 - `src/heart/runtime/render_pipeline.py`
-- `src/heart/runtime/rendering/surface_provider.py`
-- `src/heart/runtime/rendering/surface_merge.py`
 - `docs/code_flow.md`
